Prepared for first post!
I am so excited to finally get this off the ground after years of thinking about it. It just came to me after almost two years of IT cert studying and hearing about home labs—I bought a couple $7 routers from the thrift store and realized I have no clue what to do. That's when I decided I need to document for other newcomers. Trailblaze. Darn it's too late to change the blog name to that.